cloud_functions/clear_listings_cache/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.9.0